The Geotherm package

The Geotherm package

The Geotherm.Geometry module

The Geotherm.Interpolate module

The Geotherm.Integrate module

Geotherm.generate_trace — Function

generate_trace(geothermal_gradient::AbstractDimMatrix, p0::Point2D, h=0.01, n=1000)

Generate a trace of Points specifiying pressure and temperature.

The Geotherm.Geometry module

Geotherm.Geometry.Point2D — Type

Point2D(x, y)

Generate a two-dimensional point.

Geotherm.Geometry.Point3D — Type

Point3D(x, y, z)

Generate a three-dimensional point.

Geotherm.Geometry.Rectangle — Type

Rectangle(lx, rx, ly, uy)

Generate a rectangle from its four corner points.

Geotherm.Geometry.vertices — Function

vertices(rec::Rectangle)

Return four corner points of a Rectangle.

Base.in — Function

Base.in(p::Point2D, rec::Rectangle)

Test if point p is in rectangle rec.

The Geotherm.Interpolate module

Geotherm.Interpolate.linear_interpolate — Function

linear_interpolate(a, b)

Return a function of the linear interpolation between any 2 points (x₁, f(x₁)) and (x₂, g(x₂)).

Geotherm.Interpolate.bilinear_interpolate — Function

bilinear_interpolate(q₁₁, q₁₂, q₂₁, q₂₂)

If 4 points consisting of a rectangle and their z-coordinates are known, return a bilinear interpolation of the 4 points.

The Geotherm.Integrate module

Geotherm.Integrate.runge_kutta — Function

runge_kutta(p0::Point2D, f, h=0.01, n=1000)

Do a Runge–Kutta fourth-order integration of function f with starting point p0, step n, and step-size h.